--- src/app/(standard)/event/page.module.css
@@ -1,20 +1,0 @@
-.mapContainer {
-	max-width: 1120px;
-	margin: 0 auto;
-	padding: 3rem 1.5rem;
-	text-align: center;
-}
-
-.title {
-	font-size: 2rem;
-	font-weight: bold;
-	color: #475569;
-	margin-bottom: 1rem;
-}
-
-.text {
-	color: #64748b;
-	font-size: 1.125rem;
-	line-height: 1.8;
-	margin: 0.5rem 0;
-}
