--- src/app/(standard)/event/page.tsx
@@ -1,9 +1,0 @@
-import Pending from "@/components/Pending";
-
-export default function Page() {
-	return (
-		<>
-			<Pending />
-		</>
-	);
-}
