+++ src/app/(standard)/timetable/style.module.css
@@ -2,6 +2,8 @@
   min-height: 100vh;
   display: flex;
   flex-direction: column;
+  align-items: center;
+  padding: 2rem 1.5rem;
 }
 
 .inner {
@@ -65,7 +67,7 @@
   display: flex;
   flex-direction: column;
   position: relative;
-  margin: 40px 0;
+  margin: 40px auto;
   width: 100%;
   max-width: 600px;
 }
@@ -147,8 +149,8 @@
 
 /* レスポンシブ対応 */
 @media (max-width: 768px) {
-  .inner {
-    padding: 2rem 1rem;
+  .pageRoot {
+    padding: 1rem 1rem;
   }
 
   .timelineContent time {
